refactor: extract tool operation runtime
